Move utc_datetime.py from Parse to misc.
