+++ scripts/Parse/query.py
@@ -1,8 +1,9 @@
 import json
 import urllib
+
 from objects import Object
 from users import User
-from utc_datetime import UtcDateTime
+from misc.utc_datetime import UtcDateTime
 
 
 class Selector:
